+++ spec/integration/type/oneview_hypervisor_manager_spec.rb
@@ -4,12 +4,13 @@ type_class = Puppet::Type.type(:oneview_hypervisor_manager)
 
 def hm_config
   {
-    name:            'test_hm',
+    name: 'test_hm',
+    ensure: 'present'
     data:
       {
-        name:        '172.18.1.13',
-        username:    'dcs',
-        password:    'dcs'
+        name     => '172.18.13.11',
+        username => 'dcs',
+        password => 'dcs'
       }
   }
 end